--- server/package.json
@@ -1,33 +1,0 @@
-{
-  "name": "server",
-  "version": "1.0.0",
-  "description": "",
-  "main": "server.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "start": "node server.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/jadaquandt/refrigerator-app.git"
-  },
-  "author": "",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/jadaquandt/refrigerator-app/issues"
-  },
-  "homepage": "https://github.com/jadaquandt/refrigerator-app#readme",
-  "dependencies": {
-    "axios": "^0.19.2",
-    "bcrypt": "^5.0.0",
-    "body-parser": "^1.19.0",
-    "cors": "^2.8.5",
-    "express": "^4.17.1",
-    "jsonwebtoken": "^8.5.1",
-    "morgan": "^1.10.0",
-    "passport": "^0.4.1",
-    "passport-jwt": "^4.0.0",
-    "sequelize": "^6.3.3",
-    "session": "^0.1.0"
-  }
-}
