--- src/App.css
@@ -1,5 +1,0 @@
-#dashContainer {
-  background-image: url("/light_mode_option2.jpg");
-  background-size: cover;
-  background-repeat: no-repeat;
-}
